requirements.txt
stockfish==3.28.0
tk==0.1.0
mss==10.0.0
pillow==10.4.0
python-chess==1.11.1
